Moved Footer component to UI components folder. Updated imports
