+++ src/components/PageComponents/SendVerifyPage/SendVerifyPage.tsx
@@ -1,5 +1,5 @@
 import Navbar from "../../UI/Navbar/Navbar";
-import Footer from "../Welcome/Footer";
+import Footer from "../../UI/Footer/Footer";
 import SendVerify from "./SendVerify";
 
 const SendVerifyPage = () => {
